Fixed external + timer. M-mode must overwrite IRQ_SW to SIP instead of ORing it. Added RTC and syscall library. Turned on serial interrupt again

+++ src/devices/rtc.h
@@ -1,7 +1,7 @@
 #ifndef RTC_H
 #define RTC_H
 
-typedef unsigned long time_t;
+typedef unsigned long long time_t;
 
 time_t rtc_get_time(void);
